requirements.txt
matplotlib>=3.5.3
numpy>=1.24.3
opencv-python>=4.5.5.64
pillow>=9.5.0
scikit-learn>=1.2.2
seaborn>=0.12.2
torch>=2.0.1
torchvision>=0.15.2